+++ Source/Text/PDFAttributedText.swift
@@ -1,0 +1,28 @@
+//
+//  PDFAttributedText.swift
+//  TPPDF
+//
+//  Created by Philip Niedertscheider on 31/10/2017.
+//
+
+/**
+ Attributed text objects hold an instance of `NSAttributedString`
+
+ Use this class for advanced text drawing
+ */
+public class PDFAttributedText: PDFText, PDFJSONSerializable {
+
+    /**
+     Text which will be drawn
+     */
+    public var text: NSAttributedString
+
+    /**
+     Initializer
+
+     - parameter text: Text, which will be drawn
+     */
+    public init(text: NSAttributedString) {
+        self.text = text
+    }
+}