+++ Source/Text/PDFSimpleText+Equatable.swift
@@ -1,0 +1,21 @@
+//
+//  PDFSimpleText+Equatable.swift
+//  TPPDF
+//
+//  Created by Philip Niedertscheider on 09/11/2017.
+//
+
+extension PDFSimpleText {
+
+    public static func == (lhs: PDFSimpleText, rhs: PDFSimpleText) -> Bool {
+        if lhs.text != rhs.text {
+            return false
+        }
+
+        if lhs.spacing != rhs.spacing {
+            return false
+        }
+
+        return true
+    }
+}