+++ CHANGELOG.md
@@ -5,9 +5,22 @@
 
 **Implemented enhancements:**
 
+- Pagination styling using formats or closures.
+- First and last pagination page can be set.
+- Pages can be excluded from pagination.
+- Created class `Table` to manage tables.
+- Added extension `toTableContent` to class `String`, `NSAttributedString` and `UIImage`.
+- Created Error enum `TPPDFError` for error handling.
+- Added image rendering inside table
+- Moved all error handling outside of framework
+
 **Fixed bugs:**
 
 **Closed issues:**
+
+- Issue #5
+- Issue #15
+- Issue #16
 
 **Merged pull requests:**
 
@@ -19,13 +32,15 @@
 - Added property `landscapeSize` to `PageFormat` which returns the size as landscape
 - Split `PDFGenerator` into multiple files
 - Added Carthage support
-- 
+- Added table styling
+
 **Fixed bugs:**
 
 **Closed issues:**
 
 - Issue #2
 - Issue #4
+- Isseu #7
 
 **Merged pull requests:**